Changed the unmake information's representation from longs to an 8-field unencapsulated UnmakeRegister objects which yielded significant speed improvements. Perft(6) was below 16s. Then I also changed the move representation to a 4-field unencapsulated Move object which slowed perft down with about a second. Nonetheless, I am sticking with it, because it provides a much smoother 'interface' and the int representation poses difficulties when storing negative values. Besides, even in time-complexity the object-oriented solution will pay-off during a search or move ordering where the score 'field' has to be accessed and assigned to many times.
I believe it is the way to go, but I don't regret having experimented
with storing complex information in integers. I admittedly was wrong
thinking that in very performance sensitive applications, it is still
better to minimize the unpredictable garbage collection factor. Perf(6)
is solid 17s. I am currently profiling it with Java VisualVM to see if
there's anything worth trying to optimize or it is time to move on to
other parts of the engine.