Changed the data structures a little, so now the abstract class List defines the methods getTail and length which were previously accessible only in Queues, and it implements toArray. It might make this operation a tiny bit slower for turning my queues into arrays because of instead using the length field directly when setting the length of the array, it calls length() which just returns the length field but should cause overhead. However, this way only length() needs to be implemented by the subclasses, and even though stacks will most probably never call toArray in my engine, i find this solution so much more elegant, that im willing to sacrifice that milisecond or two.
I also added a method called toShort to Move that returns a short holding the objects from, to, and type values. It shall be useful when storing the best move in a positions transposition table entry, which is very memory sensitive.

Next i should restructure my 64 fold enums, because theres is a lot of memory to save there for the transposition table.