Implemented a killer table for the killer heuristic.
Refactored thousands of lines of code to 'centralize' commonly used
values in the Position class and allow for modularity. Data is still
stored in primitives wherever possible within reason for the sake of
performance, but these primitives are now accessed through enum types'
instance fields making the internal setup of the Position class
something of an 'object-based' architecture.