Fixed some move ordering bugs...
SEE was semantically wrong as it did not stop engaging in recaptures
even if it resulted in a worse position for the side to move. Now it's
fixed and handles promotions, too. MVV-LVA handles promotions accurately
now as well. The relative history table uses long tables now so it will
not overflow. Records have been wrongly only registered in the RHT if
the score was better than the high score. Quiescence search doesn't
handle checks at all anymore. The eval parameters are tuned. Currently
the engine and search control parameters are being tuned.