Major parallel search improvements and bug fixes.
Implemented ABDADA and fixed a bug causing the skipping of plies in case
of the premature termination of the master thread due to a failed window
in one of the slave threads. The time-to-depth speedup is solid (over
1.6 on 2 cores and over 2.1 on 4 cores), the NPS scaling is sub-optimal
(presumably due to high memory traffic). Made major modifications to the
hash table; the entries are now pre-allocated. Also did some additional
refactoring.

Next up, the evaluation needs to be fixed up and finally, the parameters
have to be retuned.